+++ src/main/scala/com/github/derghust/pekkotemplate/behavior/PekkoSupervisor.scala
@@ -3,6 +3,9 @@ package com.github.derghust.pekkotemplate.behavior
 import com.github.derghust.pekkotemplate.message.Message
 import org.apache.pekko.actor.typed.Behavior
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
+import com.github.derghust.pekkotemplate.message.ArgumentMessage
+import com.github.derghust.pekkotemplate.database.UserDB
+import com.github.derghust.pekkotemplate.database.DoobieUtil
 
 object PekkoSupervisor {
   def apply(): Behavior[Message] = Behaviors.setup(_ => empty)
@@ -10,7 +13,13 @@ object PekkoSupervisor {
   private lazy val empty: Behavior[Message] =
     Behaviors.receive[Message] { (context, message) =>
       message match
-        case unhandled =>
+        case ArgumentMessage(cmd, args) =>
+          context.log.info(s"Processing Argument message [cmd=$cmd; args=$args]")
+          cmd match
+            case unhandledCmd =>
+              context.log.error(s"Unhandled command! [cmd=$cmd]")
+          empty
+        case unhandled                  =>
           context.log.error(s"Unhandled message! [message=$unhandled]")
           empty
     }