+++ src/main/scala/com/github/derghust/pekkotemplate/message/ArgumentMessage.scala
@@ -1,0 +1,3 @@
+package com.github.derghust.pekkotemplate.message
+
+final case class ArgumentMessage(cmd: String, args: List[String]) extends Message