+++ src/main/scala/com/github/derghust/pekkotemplate/structure/Entity.scala
@@ -1,0 +1,3 @@
+package com.github.derghust.pekkotemplate.structure
+
+final case class Entity(id: Id, label: String, description: String)